requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn>=0.27.0
websockets>=12.0

# Data handling
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Date handling
python-dateutil>=2.8.0

# Vector Store / RAG
numpy>=1.26.0
faiss-cpu>=1.7.0

# Async support
aiofiles>=23.0.0

# HTTP client (用于调用 LLM API)
httpx>=0.27.0

# JSON handling
orjson>=3.9.0
